+++ ppa/common/tests.py
@@ -1,15 +1,13 @@
-from datetime import datetime
 from unittest.mock import Mock
 
-from django.contrib.auth.models import User, Group
-from django.test import override_settings, RequestFactory, TestCase
+from django.contrib.auth.models import Group, User
+from django.test import RequestFactory, TestCase, override_settings
 from django.urls import reverse
-from wagtail.core.models import Site, Page
+from wagtail.core.models import Page, Site
 
-from ppa.common.admin import LocalUserAdmin
-from ppa.common.views import VaryOnHeadersMixin, AjaxTemplateMixin, \
-    LastModifiedMixin
 from ppa.archive.views import DigitizedWorkListView
+from ppa.common.admin import LocalUserAdmin
+from ppa.common.views import AjaxTemplateMixin, VaryOnHeadersMixin
 
 
 class TestLocalUserAdmin(TestCase):
@@ -150,20 +148,3 @@ class TestAnalytics(TestCase):
         )
         # should not have the call to gtags snippet
         self.assertNotContains(res, 'gtag(')
-
-
-
-
-
-class TestLastModifiedMixin(TestCase):
-
-    def test_solr_timestamp_to_datetime(self):
-        # with microseconds
-        solr_dt = LastModifiedMixin.solr_timestamp_to_datetime('2018-07-02T21:08:46.428Z')
-        assert solr_dt == datetime(2018, 7, 2, 21, 8, 46)
-        # without
-        solr_dt = LastModifiedMixin.solr_timestamp_to_datetime('2018-07-02T21:08:46Z')
-        assert solr_dt == datetime(2018, 7, 2, 21, 8, 46)
-
-
-